// react-frontend/src/api/models/financial-period-statistic.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Spys API
 * REST API for Spys
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface FinancialPeriodStatistic
 */
export interface FinancialPeriodStatistic {
    /**
     * 
     * @type {string}
     * @memberof FinancialPeriodStatistic
     */
    'periodStart'?: string;
    /**
     * 
     * @type {string}
     * @memberof FinancialPeriodStatistic
     */
    'periodEnd'?: string;
    /**
     * 
     * @type {number}
     * @memberof FinancialPeriodStatistic
     */
    'totalRevenue'?: number;
    /**
     * 
     * @type {number}
     * @memberof FinancialPeriodStatistic
     */
    'totalRevenuePending'?: number;
    /**
     * 
     * @type {number}
     * @memberof FinancialPeriodStatistic
     */
    'totalRefunds'?: number;
    /**
     * 
     * @type {number}
     * @memberof FinancialPeriodStatistic
     */
    'totalCreditTransactions'?: number;
}